third_party/blink/web_tests/fast/dom/getElementById-consistency4-expected.txt

Test that DOMSubtreeModified listeners cannot cause TreeOrderedMap to be incorrect

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.contains(el) is true
PASS document.getElementById('el') is null
PASS successfullyParsed is true

TEST COMPLETE
